refactor: migrate project to feature architecture
