+++ package.json
@@ -9,6 +9,7 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@react-native-async-storage/async-storage": "2.2.0",
     "@react-navigation/native": "^7.1.19",
     "@react-navigation/stack": "^7.6.3",
     "expo": "~54.0.23",
