+++ src/features/auth/types/auth.type.ts
@@ -1,0 +1,40 @@
+import { Level } from '@/shared/types/common';
+import { CoreEntity } from '@/shared/types/common/cores.type';
+export * from './auth.type';
+export interface AuthState {
+  accessToken: string | null;
+  user: User | null;
+  isAuthenticated: boolean;
+  idToken: string | null;
+  refreshToken: string | null;
+  signInStatus: (authenticated: Authenticated) => Promise<void>;
+  signOut: () => Promise<void>;
+  restoreSession: () => Promise<void>;
+}
+
+export interface User extends CoreEntity {
+  avatar: Object;
+  username: string;
+  email: string;
+  score: number;
+  userRole: UserRoles;
+  level: Level;
+}
+
+export enum UserRoles {
+  PLAYER = 'PLAYER',
+  ADMIN = 'ADMIN',
+}
+
+export interface Authenticated {
+  accessToken: string;
+  idToken: string;
+  refreshToken: string;
+  expiresIn: number;
+}
+
+export interface SignUpDto {
+  email: string;
+  username: string;
+  password: string;
+}
