--- src/question/services/question.service.ts
@@ -1,38 +1,0 @@
-import { ApiResponse } from '@/shared/api/types';
-import { apiService } from '../../shared/api/api.service';
-import { API_ENDPOINTS } from '../../shared/api/apiConfig';
-import { CreateQuestionDto, Question } from '../types';
-
-export interface UpdateQuestionDto extends Partial<CreateQuestionDto> {}
-
-export const questionService = {
-  // Obtener todas las preguntas
-  getAll: async (): Promise<ApiResponse<Question[]>> => {
-    return apiService.get<Question[]>(API_ENDPOINTS.QUESTIONS);
-  },
-
-  // Obtener una pregunta por ID
-  getById: async (id: string): Promise<ApiResponse<Question>> => {
-    return apiService.get<Question>(`${API_ENDPOINTS.QUESTIONS}/${id}`);
-  },
-
-  // Crear una nueva pregunta
-  create: async (data: CreateQuestionDto): Promise<ApiResponse<Question>> => {
-    return apiService.post<Question>(API_ENDPOINTS.QUESTIONS, data);
-  },
-
-  // Crear múltiples preguntas
-  createMany: async (data: CreateQuestionDto[]): Promise<ApiResponse<Question[]>> => {
-    return apiService.post<Question[]>(`${API_ENDPOINTS.QUESTIONS}/bulk`, data);
-  },
-
-  // Actualizar una pregunta
-  update: async (id: string, data: UpdateQuestionDto): Promise<ApiResponse<Question>> => {
-    return apiService.patch<Question>(`${API_ENDPOINTS.QUESTIONS}/${id}`, data);
-  },
-
-  // Eliminar una pregunta
-  delete: async (id: string): Promise<ApiResponse<void>> => {
-    return apiService.delete<void>(`${API_ENDPOINTS.QUESTIONS}/${id}`);
-  },
-};
