+++ src/shared/adapters/state.adapter.ts
