+++ src/shared/ui/index.ts
@@ -1,0 +1,4 @@
+﻿export { default as Button } from '@/components/Button.component';
+export { default as Input } from '@/components/Input.component';
+
+export { default as FilterSection } from '@/components/FilterSection.component';
